requirements.txt
Pillow==4.3.0
keras==2.1.1
numpy==1.13.3
